# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Production ERP Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all pages (menu structure)
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page/menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
  /pages/{id}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get page by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    put:
      operationId: updatePage
      tags: [pages]
      summary: Update page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /pages/reorder:
    post:
      operationId: reorderPages
      tags: [pages]
      summary: Reorder pages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # FIELDS
  /entities/{entityId}/fields:
    get:
      operationId: listEntityFields
      tags: [fields]
      summary: List fields for an entity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Field"
    post:
      operationId: createEntityField
      tags: [fields]
      summary: Create a field on an entity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldInput"
      responses:
        "201":
          description: Field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
  /fields/{id}:
    get:
      operationId: getField
      tags: [fields]
      summary: Get field by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Field found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
    put:
      operationId: updateField
      tags: [fields]
      summary: Update field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldUpdate"
      responses:
        "200":
          description: Field updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
    delete:
      operationId: deleteField
      tags: [fields]
      summary: Delete field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /fields/reorder:
    post:
      operationId: reorderFields
      tags: [fields]
      summary: Reorder fields within an entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldsReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # STATUSES
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Multilingual label stored as JSONB {ru: ..., en: ..., he: ...}
    MultilingualText:
      type: object
      properties:
        ru:
          type: string
        en:
          type: string
        he:
          type: string

    # AUTH
...
    Page:
      type: object
      required: [id, nameJson, icon, sortOrder, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        path:
          type: ["string", "null"]
        parentPageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        children:
          type: array
          items:
            $ref: "#/components/schemas/Page"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PageInput:
      type: object
      required: [nameJson, icon]
      properties:
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        path:
          type: ["string", "null"]
        parentPageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
          default: true
    PageUpdate:
      type: object
      properties:
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        path:
          type: ["string", "null"]
        parentPageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
    ReorderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, sortOrder]
            properties:
              id:
                type: integer
              sortOrder:
                type: integer

    # ENTITIES
...
    FieldType:
      type: string
      enum:
        - text
        - textarea
        - number
        - boolean
        - date
        - datetime
        - select
        - email
        - url
        - phone
    Field:
      type: object
      required: [id, entityId, fieldKey, nameJson, fieldType, isRequired, optionsJson, sortOrder, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        entityId:
          type: integer
        fieldKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        fieldType:
          $ref: "#/components/schemas/FieldType"
        isRequired:
          type: boolean
        defaultValue:
          type: ["string", "null"]
        optionsJson:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FieldInput:
      type: object
      required: [fieldKey, nameJson, fieldType]
      properties:
        fieldKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        fieldType:
          $ref: "#/components/schemas/FieldType"
        isRequired:
          type: boolean
          default: false
        defaultValue:
          type: ["string", "null"]
        optionsJson:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
          default: true
    FieldsReorderInput:
      type: object
      required: [entityId, items]
      properties:
        entityId:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [id, sortOrder]
            properties:
              id:
                type: integer
              sortOrder:
                type: integer
    FieldUpdate:
      type: object
      properties:
        fieldKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        fieldType:
          $ref: "#/components/schemas/FieldType"
        isRequired:
          type: boolean
        defaultValue:
          type: ["string", "null"]
        optionsJson:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
...